Moved LibraryCustomerBuilder to separate package
